\name{Metropolis Hastings Algorithm-package}
\alias{Metropolis Hastings Algorithm-package}
\alias{Metropolis Hastings Algorithm}
\docType{package}
\title{
\packageTitle{Metropolis Hastings Algorithm}
}
\description{
\packageDescription{Metropolis Hastings Algorithm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Metropolis Hastings Algorithm}
\packageIndices{Metropolis Hastings Algorithm}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Metropolis Hastings Algorithm}

Maintainer: \packageMaintainer{Metropolis Hastings Algorithm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
